+++ src/analytics/Properties.tsx
@@ -54,8 +54,7 @@ import { ErrorMessages } from 'src/app/ErrorMessages'
 import { AddAssetsActionType } from 'src/components/AddAssetsBottomSheet'
 import { TokenPickerOrigin } from 'src/components/TokenBottomSheet'
 import { DappSection } from 'src/dapps/types'
-import { BeforeDepositActionName } from 'src/earn/EarnPoolInfoScreen'
-import { SerializableRewardsInfo } from 'src/earn/types'
+import { BeforeDepositActionName, SerializableRewardsInfo } from 'src/earn/types'
 import { ProviderSelectionAnalyticsData } from 'src/fiatExchanges/types'
 import { CICOFlow, FiatExchangeFlow, PaymentMethod } from 'src/fiatExchanges/utils'
 import { HomeActionName, NotificationBannerCTATypes, NotificationType } from 'src/home/types'
@@ -1579,7 +1578,7 @@ interface EarnEventsProperties {
     depositTokenId: string
   }
   [EarnEvents.earn_entrypoint_press]: undefined
-  [EarnEvents.earn_add_crypto_action_press]: {
+  [EarnEvents.earn_before_deposit_action_press]: {
     action: BeforeDepositActionName
   } & TokenProperties
   [EarnEvents.earn_deposit_provider_info_press]: EarnDepositProperties
